create account command working with start transaction and database atomic transaction

+++ Services/Core/CategoryService.cs
@@ -10,10 +10,13 @@ namespace FinancialAdvisorTelegramBot.Services.Core
         private const string DefaultCategoryName = "Default";
 
         private readonly ICategoryRepository _repository;
+        private readonly ITransactionRepository _transactionRepository;
 
-        public CategoryService(ICategoryRepository repository)
+
+        public CategoryService(ICategoryRepository repository, ITransactionRepository transactionRepository)
         {
             _repository = repository;
+            _transactionRepository = transactionRepository;
         }
 
         public async Task<Category> CreateCategory(int userId, string name, string? description)
@@ -25,20 +28,24 @@ namespace FinancialAdvisorTelegramBot.Services.Core
             {
                 Name = name,
                 Description = description,
-                UserId = userId
+                UserId = userId,
+                CreatedAt = DateTime.Now,
             };
             Category added = await _repository.Add(category);
             return await _repository.GetById(added.Id) 
-                ?? throw new Exception("Category was not added");
+                ?? throw new Exception("Category was not created");
         }
 
         public async Task DeleteCategory(Category category)
         {
+            if (await _transactionRepository.HasAnyTransactionByCategory(category.Id))
+                throw new Exception("Cannot delete category with transactions");
             await _repository.Delete(category);
         }
 
         public async Task<Category> UpdateCategory(Category category)
         {
+            category.UpdatedAt = DateTime.Now;
             return await _repository.Update(category);
         }
 
@@ -47,7 +54,7 @@ namespace FinancialAdvisorTelegramBot.Services.Core
             return await _repository.GetCategoriesByUser(userId);
         }
 
-        public async Task<Category> GetOtherwiseCreateDefaultCategory(int userId)
+        public async Task<Category> GetOrOtherwiseCreateDefaultCategory(int userId)
         {
             return await _repository.GetCategoryByName(userId, DefaultCategoryName) 
                 ?? await CreateCategory(userId, DefaultCategoryName, null);
